global-hotkeys/requirements.txt
dbus-python==1.3.2
libevdev==0.11
python-libinput @ git+https://github.com/OzymandiasTheGreat/python-libinput.git@e5bdb2f5d6cc67452f7af32ecf9bcce4e6f2d8d8
